[FEAT] filter에 넣었던 로직을 컨트롤러에서 하도록 수정

+++ src/main/java/com/jdong/studycafe/favorites/controller/FavoriteController.java
@@ -4,6 +4,8 @@ import com.jdong.studycafe.config.auth.CustomUserDetails;
 import com.jdong.studycafe.favorites.dto.FavoriteDTO;
 import com.jdong.studycafe.favorites.dto.FavoriteRequestDTO;
 import com.jdong.studycafe.favorites.service.FavoriteService;
+import com.jdong.studycafe.orders.exception.IsStudyingException;
+import com.jdong.studycafe.orders.service.StudyService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.http.ResponseEntity;
@@ -21,11 +23,16 @@ import java.util.List;
 public class FavoriteController {
 
     private final FavoriteService favoriteService;
+    private final StudyService studyService;
 
     @GetMapping("")
     public ResponseEntity<HashMap<String, Object>> getFavoriteListByMemberId(
             Authentication authentication) {
         CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        Boolean isStudying = studyService.isStudying(userDetails.getMember().getId());
+        if (isStudying == Boolean.TRUE) {
+            throw new IsStudyingException(userDetails.getMember().getId().toString());
+        }
         Long memberId = userDetails.getMember().getId();
         List<FavoriteDTO> favoriteList = favoriteService.getFavoriteListByMemberId(memberId);
 
@@ -42,6 +49,10 @@ public class FavoriteController {
             Authentication authentication) {
 
         CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        Boolean isStudying = studyService.isStudying(userDetails.getMember().getId());
+        if (isStudying == Boolean.TRUE) {
+            throw new IsStudyingException(userDetails.getMember().getId().toString());
+        }
         FavoriteDTO favoriteDTO = favoriteService.addFavorite(favoriteRequestDTO, userDetails.getMember().getId());
 
         HashMap<String, Object> result = new HashMap<>();
@@ -57,6 +68,10 @@ public class FavoriteController {
             Authentication authentication
     ) {
         CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        Boolean isStudying = studyService.isStudying(userDetails.getMember().getId());
+        if (isStudying == Boolean.TRUE) {
+            throw new IsStudyingException(userDetails.getMember().getId().toString());
+        }
 
         HashMap<String, Object> result = favoriteService.deleteFavorite(favoriteId, userDetails.getMember().getId());
